@@ -19,13 +19,14 @@
     "test:e2e": "export NODE_ENV=test-e2e && jest --config ./tests/jest-e2e.json --detectOpenHandles --forceExit --verbose"
   },
   "dependencies": {
-    "@nestjs/common": "^7.6.17",
-    "@nestjs/core": "^7.6.17",
-    "@nestjs/cqrs": "^7.0.1",
-    "@nestjs/microservices": "^7.6.17",
-    "@nestjs/platform-express": "^7.6.17",
-    "@nestjs/swagger": "^4.8.0",
-    "@nestjs/terminus": "^7.2.0",
+    "@nestjs/axios": "^0.0.1",
+    "@nestjs/common": "^8.0.4",
+    "@nestjs/core": "^8.0.4",
+    "@nestjs/cqrs": "^8.0.0",
+    "@nestjs/microservices": "^8.0.4",
+    "@nestjs/platform-express": "^8.0.4",
+    "@nestjs/swagger": "^5.0.4",
+    "@nestjs/terminus": "^8.0.0-next.0",
     "@simplewebauthn/server": "^3.0.0",
     "agenda": "^4.1.3",
     "class-transformer": "^0.4.0",
@@ -34,15 +35,15 @@
     "cookie-parser": "^1.4.5",
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
-    "express-rate-limit": "^5.2.6",
+    "express-rate-limit": "^5.3.0",
     "express-session": "^1.17.2",
     "hbs": "^4.1.2",
     "helmet": "^4.6.0",
     "i18n": "^0.13.3",
-    "joi": "^17.4.0",
-    "mongodb": "^3.6.9",
-    "mongoose": "^5.12.12",
-    "mqtt": "^4.2.6",
+    "joi": "^17.4.1",
+    "mongodb": "^3.6.10",
+    "mongoose": "^5.13.3",
+    "mqtt": "^4.2.8",
     "node-jose": "^2.0.0",
     "oauth2orize": "^1.11.0",
     "oauth2orize-openid": "^0.4.0",
@@ -56,38 +57,36 @@
     "qrcode": "^1.4.4",
     "rate-limit-mongo": "^2.3.1",
     "reflect-metadata": "^0.1.13",
-    "rxjs": "^6.6.7",
+    "rxjs": "^7.2.0",
     "saslprep": "^1.0.3",
     "speakeasy": "^2.0.0",
     "swagger-ui-express": "^4.1.6",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
-    "@nestjs/cli": "^7.6.0",
-    "@nestjs/testing": "^7.6.17",
+    "@nestjs/cli": "^8.0.2",
+    "@nestjs/testing": "^8.0.4",
     "@simplewebauthn/typescript-types": "^3.0.0",
-    "@types/express": "^4.17.12",
-    "@types/express-session": "^1.17.3",
-    "@types/jest": "^26.0.23",
-    "@types/joi": "^17.2.2",
-    "@types/mongodb": "^3.6.17",
+    "@types/express": "^4.17.13",
+    "@types/express-session": "^1.17.4",
+    "@types/jest": "^26.0.24",
+    "@types/mongodb": "^3.6.20",
     "@types/node": "^15.6.1",
-    "@types/oauth": "^0.9.1",
-    "@types/passport-local": "^1.0.33",
+    "@types/passport-local": "^1.0.34",
     "@types/supertest": "^2.0.11",
-    "@typescript-eslint/eslint-plugin": "^4.25.0",
-    "@typescript-eslint/parser": "^4.25.0",
-    "eslint": "^7.27.0",
-    "eslint-plugin-jsdoc": "^35.0.0",
+    "@typescript-eslint/eslint-plugin": "^4.28.3",
+    "@typescript-eslint/parser": "^4.28.3",
+    "eslint": "^7.30.0",
+    "eslint-plugin-jsdoc": "^35.4.3",
     "eslint-plugin-prefer-arrow": "^1.2.3",
-    "jest": "^27.0.3",
-    "prettier": "^2.3.0",
+    "jest": "^27.0.6",
+    "prettier": "^2.3.2",
     "supertest": "^6.1.3",
-    "ts-jest": "^27.0.1",
-    "ts-loader": "^9.2.2",
-    "ts-node": "^10.0.0",
-    "tsconfig-paths": "^3.9.0",
-    "typescript": "^4.3.2"
+    "ts-jest": "^27.0.3",
+    "ts-loader": "^9.2.3",
+    "ts-node": "^10.1.0",
+    "tsconfig-paths": "^3.10.1",
+    "typescript": "^4.3.5"
   },
   "jest": {
     "testEnvironment": "node",
