@@ -13,48 +13,48 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.0.2",
-    "@angular/cdk": "^12.0.2",
-    "@angular/common": "~12.0.2",
-    "@angular/compiler": "~12.0.2",
-    "@angular/core": "~12.0.2",
+    "@angular/animations": "~12.1.2",
+    "@angular/cdk": "^12.1.2",
+    "@angular/common": "~12.1.2",
+    "@angular/compiler": "~12.1.2",
+    "@angular/core": "~12.1.2",
     "@angular/flex-layout": "^12.0.0-beta.34",
-    "@angular/forms": "~12.0.2",
-    "@angular/material": "^12.0.2",
-    "@angular/platform-browser": "~12.0.2",
-    "@angular/platform-browser-dynamic": "~12.0.2",
-    "@angular/router": "~12.0.2",
-    "angular-oauth2-oidc": "^10.0.3",
-    "angular-oauth2-oidc-jwks": "^9.0.0",
+    "@angular/forms": "~12.1.2",
+    "@angular/material": "^12.1.2",
+    "@angular/platform-browser": "~12.1.2",
+    "@angular/platform-browser-dynamic": "~12.1.2",
+    "@angular/router": "~12.1.2",
+    "angular-oauth2-oidc": "^12.0.0",
+    "angular-oauth2-oidc-jwks": "^12.0.0",
     "hammerjs": "^2.0.8",
     "rxjs": "~6.6.7",
     "tslib": "^2.2.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.0.2",
-    "@angular/cli": "~12.0.2",
-    "@angular/compiler-cli": "~12.0.2",
-    "@angular/language-service": "^12.0.2",
+    "@angular-devkit/build-angular": "~12.1.2",
+    "@angular/cli": "~12.1.2",
+    "@angular/compiler-cli": "~12.1.2",
+    "@angular/language-service": "^12.1.2",
     "@types/jasmine": "^3.7.6",
     "@types/node": "^15.6.1",
-    "@typescript-eslint/eslint-plugin": "^4.25.0",
-    "@typescript-eslint/parser": "^4.25.0",
+    "@typescript-eslint/eslint-plugin": "^4.28.3",
+    "@typescript-eslint/parser": "^4.28.3",
     "codelyzer": "^6.0.2",
     "eslint": "^7.27.0",
-    "eslint-plugin-jsdoc": "^35.0.0",
+    "eslint-plugin-jsdoc": "^35.4.3",
     "eslint-plugin-prefer-arrow": "^1.2.3",
-    "jasmine-core": "~3.7.1",
+    "jasmine-core": "~3.8.0",
     "jasmine-spec-reporter": "~7.0.0",
     "karma": "~6.3.2",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.0.3",
     "karma-coverage-istanbul-reporter": "^3.0.3",
     "karma-jasmine": "~4.0.1",
-    "karma-jasmine-html-reporter": "^1.6.0",
+    "karma-jasmine-html-reporter": "^1.7.0",
     "prettier": "^2.3.0",
     "protractor": "~7.0.0",
-    "ts-node": "~10.0.0",
-    "typescript": "~4.2.4"
+    "ts-node": "~10.1.0",
+    "typescript": "~4.3.5"
   }
 }
