@@ -1,6 +1,7 @@
-import { CommandHandler, ICommandHandler, EventPublisher } from '@nestjs/cqrs';
-import { UploadNewAvatarCommand } from './upload-new-avatar.command';
+import { CommandHandler, EventPublisher, ICommandHandler } from '@nestjs/cqrs';
+import { lastValueFrom } from 'rxjs';
 import { ProfileManagerAggregateService } from '../../../profile-management/aggregates/profile-manager-aggregate/profile-manager-aggregate.service';
+import { UploadNewAvatarCommand } from './upload-new-avatar.command';
 
 @CommandHandler(UploadNewAvatarCommand)
 export class UploadNewAvatarHandler
@@ -15,9 +16,9 @@ export class UploadNewAvatarHandler
     const { avatarFile, clientHttpRequest: clientHttpRequest } = command;
 
     const aggregate = this.publisher.mergeObjectContext(this.manager);
-    await this.manager
-      .uploadNewAvatar(avatarFile, clientHttpRequest)
-      .toPromise();
+    await lastValueFrom(
+      this.manager.uploadNewAvatar(avatarFile, clientHttpRequest),
+    );
     aggregate.commit();
   }
 }
