@@ -1,13 +1,12 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { HttpModule } from '@nestjs/common';
+import { HttpModule } from '@nestjs/axios';
 import { CqrsModule } from '@nestjs/cqrs';
-
-import { OAuth2Service } from './oauth2.service';
-import { UserService } from '../../../user-management/entities/user/user.service';
+import { Test, TestingModule } from '@nestjs/testing';
 import { BearerTokenService } from '../../../auth/entities/bearer-token/bearer-token.service';
-import { ServerSettingsService } from '../../../system-settings/entities/server-settings/server-settings.service';
 import { ClientService } from '../../../client-management/entities/client/client.service';
+import { ServerSettingsService } from '../../../system-settings/entities/server-settings/server-settings.service';
+import { UserService } from '../../../user-management/entities/user/user.service';
 import { UserClaimService } from '../../entities/user-claim/user-claim.service';
+import { OAuth2Service } from './oauth2.service';
 
 describe('OAuth2Service', () => {
   let service: OAuth2Service;
