@@ -1,4 +1,5 @@
-import { Module, Global, HttpModule } from '@nestjs/common';
+import { Module, Global } from '@nestjs/common';
+import { HttpModule } from '@nestjs/axios';
 import { AuthEntitiesModule } from './entities/entities.module';
 import { AuthServerVerificationGuard } from './guards/authserver-verification.guard';
 import { RoleGuard } from './guards/role.guard';
