@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { HttpService } from '@nestjs/common';
+import { HttpService } from '@nestjs/axios';
 import { from } from 'rxjs';
 import { SettingsService } from './settings.service';
 import { ServerSettingsService } from '../../entities/server-settings/server-settings.service';
