@@ -1,13 +1,14 @@
-import { Injectable, BadRequestException } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { AggregateRoot } from '@nestjs/cqrs';
-import { UploadAvatarMetaDataService } from '../../../profile-management/policies/upload-avatar-meta-data/upload-avatar-meta-data.service';
-import { NewAvatarUploadedEvent } from '../../../profile-management/events/new-avatar-uploaded/new-avatar-uploaded.event';
+import { lastValueFrom } from 'rxjs';
 import { map, switchMap } from 'rxjs/operators';
-import { ProfileService } from '../../entities/profile/profile.service';
+import { ClientTokenManagerService } from '../../../auth/aggregates/client-token-manager/client-token-manager.service';
 import { INVALID_USER } from '../../../constants/messages';
+import { NewAvatarUploadedEvent } from '../../../profile-management/events/new-avatar-uploaded/new-avatar-uploaded.event';
+import { UploadAvatarMetaDataService } from '../../../profile-management/policies/upload-avatar-meta-data/upload-avatar-meta-data.service';
+import { ProfileService } from '../../entities/profile/profile.service';
 import { AvatarDeletedEvent } from '../../events/avatar-deleted/avatar-deleted.event';
 import { ProfileDeletedEvent } from '../../events/profile-deleted/profile-deleted.event';
-import { ClientTokenManagerService } from '../../../auth/aggregates/client-token-manager/client-token-manager.service';
 
 @Injectable()
 export class ProfileManagerAggregateService extends AggregateRoot {
@@ -40,24 +41,22 @@ export class ProfileManagerAggregateService extends AggregateRoot {
   async deleteAvatar(uuid: string, req) {
     const profile = await this.profile.findOne({ uuid });
     if (!profile) throw new BadRequestException(INVALID_USER);
-    return this.avatarMetaData
-      .uploadNewAvatarMetaData(req.token.accessToken)
-      .pipe(
+    return lastValueFrom(
+      this.avatarMetaData.uploadNewAvatarMetaData(req.token.accessToken).pipe(
         map(bucketSettings => {
           if (bucketSettings) {
             this.apply(new AvatarDeletedEvent(profile, bucketSettings));
           }
         }),
-      )
-      .toPromise();
+      ),
+    );
   }
 
   async deleteProfile(uuid: string) {
     const profile = await this.profile.findOne({ uuid });
     if (profile) {
-      await this.clientToken
-        .getClientToken()
-        .pipe(
+      await lastValueFrom(
+        this.clientToken.getClientToken().pipe(
           switchMap(token => {
             return this.avatarMetaData.uploadNewAvatarMetaData(
               token.accessToken,
@@ -69,8 +68,8 @@ export class ProfileManagerAggregateService extends AggregateRoot {
             }
             this.apply(new ProfileDeletedEvent(profile));
           }),
-        )
-        .toPromise();
+        ),
+      );
     }
   }
 }
