@@ -2,7 +2,8 @@ import { IEventHandler, EventsHandler } from '@nestjs/cqrs';
 import { throwError } from 'rxjs';
 import { switchMap, catchError, retry } from 'rxjs/operators';
 import { AvatarDeletedEvent } from './avatar-deleted.event';
-import { HttpService, ForbiddenException } from '@nestjs/common';
+import { ForbiddenException } from '@nestjs/common';
+import { HttpService } from '@nestjs/axios';
 import { ClientTokenManagerService } from '../../../auth/aggregates/client-token-manager/client-token-manager.service';
 import { BEARER, CONTENT_TYPE, APP_JSON } from '../../../constants/app-strings';
 import { INVALID_TOKEN } from '../../../constants/messages';
