@@ -1,4 +1,5 @@
-import { Injectable, HttpService } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
+import { HttpService } from '@nestjs/axios';
 import { from, of } from 'rxjs';
 import { ServerSettingsService } from '../../../system-settings/entities/server-settings/server-settings.service';
 import { switchMap, map } from 'rxjs/operators';
