@@ -1,9 +1,10 @@
+import { HttpService } from '@nestjs/axios';
 import {
-  Injectable,
-  HttpService,
   BadRequestException,
+  Injectable,
   NotFoundException,
 } from '@nestjs/common';
+import { lastValueFrom } from 'rxjs';
 import { TokenCache } from '../../../auth/entities/token-cache/token-cache.entity';
 import { ServerSettingsService } from '../../../system-settings/entities/server-settings/server-settings.service';
 
@@ -23,11 +24,11 @@ export class ClientIdMustExistOnAuthorizationServerPolicyService {
     const requestUrl =
       settings.authServerURL + GET_CLIENT_ENDPOINT + '/' + clientId;
     try {
-      const response = await this.http
-        .get(requestUrl, {
+      const response = await lastValueFrom(
+        this.http.get(requestUrl, {
           headers: { Authorization: 'Bearer ' + token.accessToken },
-        })
-        .toPromise();
+        }),
+      );
       if (response.data.clientId !== clientId) {
         throw new NotFoundException({ clientId });
       }
