@@ -1,4 +1,5 @@
-import { Module, HttpModule } from '@nestjs/common';
+import { Module } from '@nestjs/common';
+import { HttpModule } from '@nestjs/axios';
 import { ServiceManagementControllers } from './controllers';
 import { ServiceManagementAggregates } from './aggregates';
 import { CqrsModule } from '@nestjs/cqrs';
